+++ Explorer/Assets/Plugins/RustSegment/rust-segment/src/cabi.rs
@@ -96,7 +96,7 @@ pub extern "C" fn segment_server_flush() -> OperationHandleId {
 }
 
 #[no_mangle]
-pub extern "C" fn segment_server_unflushed_batches_count() -> usize {
+pub extern "C" fn segment_server_unflushed_batches_count() -> u64 {
     SEGMENT_SERVER.unflushed_batches_count()
 }
